Fixed bug allowing sensor to be open very quickly and not trigger alarm
There is still a very short windows, but I don't think it really matters
This could be addressed on the server side. If the server gets a closed
update when the station was alrady closed and the cause is not a timer,
it could be that the alarm was triggered. I need to test it to see if
it always means that.